requirements.txt
black
flake8
Pillow
httpx
sanic
